sofa-summer-25-250812-working-ui/data-processing/pyproject.toml
[project]
name = "sofa"
version = "1.0.0"
description = "Security Orchestration for Apple - Feed builder and processor"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Mac Admins"},
]

dependencies = [
    "requests>=2.31.0",
    "httpx[http2]>=0.27.0",
    "packaging>=24.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.2",
    "feedgen>=1.0.0",
    "python-dateutil>=2.9.0",
    "pydantic>=2.8.2",
    "certifi",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.5.0",
]

[project.scripts]
build-feeds = "scripts.build_feeds:main"
postprocess-feeds = "scripts.postprocess_feeds:main"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["scripts"]
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]